+++ CHANGELOG.md
@@ -33,6 +33,7 @@ Unless otherwise specified, any version comparison below is the comparison of th
 - (API) Added APIs for setting the scale of entity:
   - Added methods `Entity.setScale()` and `Entity.getScale()`.
   - Added method `Entity.getBaseAABB()` to get the original aabb which is not affected by the scale factor.
+- (API) Added method `EntityPhysicsComponent.resetFallDistance()`.
 - Added properties `network-settings.raknet-send-cookie` and `network-settings.raknet-max-mtu` to `server.properties`.
 - Implemented Totem of Undying.
 - Implemented End Crystal.